// ui/coreShared/ts/types/contracts.d.ts
export type MarketType = 'binary' | 'categorical' | 'scalar';
// ...
export type QuestionData = {
    title: string;
    description: string;
    startTime: bigint;
    endTime: bigint;
    numTicks: bigint;
    displayValueMin: bigint;
    displayValueMax: bigint;
    answerUnit: string;
};
export type ZoltarChildUniverseSummary = {
    exists: boolean;
    forkTime: bigint;
    outcomeIndex: bigint;
    outcomeLabel: string;
    parentUniverseId: bigint;
    reputationToken: Address;
    universeId: bigint;
};
export type ZoltarUniverseSummary = {
    childUniverses: ZoltarChildUniverseSummary[];
    forkBurnDivisor?: bigint;
    forkThresholdAttoRep: bigint;
    forkQuestionDetails: MarketDetails | undefined;
    forkTime: bigint;
    forkingOutcomeIndex: bigint;
    hasForked: boolean;
    parentUniverseId: bigint;
    reputationToken: Address;
    totalTheoreticalSupplyAttoRep: bigint;
    universeId: bigint;
    zoltarAddress?: Address;
};
// ...
export type MarketDetails = QuestionData & {
    createdAt: bigint;
    exists: boolean;
    marketType: MarketType;
    outcomeLabels: string[];
    questionId: string;
};